Add the links button to the app header and the pr/issue header. Style tweaks.
