+++ source/content-styles.css
@@ -1,5 +1,24 @@
+.copy-github-link-container div.dropdown-menu {
+    top:6px;
+    width: max-content;
+    max-width: calc(100vw - 320px);
+}
+.copy-github-link-container .dropdown-menu div {
+    margin: 12px;
+}
+.copy-github-link-container h5 {
+    font-weight: normal;
+}
+.copy-github-link-body {
+  background-position: center;
+  background-repeat: no-repeat;
+}
+.copy-github-link-body div {
+  background-color: rgba(255, 255, 255, 0.9);
+  padding: 0;
+}
 UL.copy-github-link-list {
-    padding: 10px 20px;
+    padding: 10px 0px 10px 20px;
     list-style-type: none;
 }
 UL.copy-github-link-list LI {
